requirements.txt
pandas==1.3.3
matplotlib==3.4.3
numpy==1.21.2
openpyxl==3.0.9
mysql-connector-python==8.0.27
